trigger price handled as string, same as price
