+++ core/src/main/java/bisq/core/util/validation/AltcoinValidator.java
@@ -15,24 +15,23 @@
  * along with Haveno. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.util.validation;
+package bisq.core.util.validation;
 
 import javax.inject.Inject;
 
-public class FiatPriceValidator extends MonetaryValidator {
+public class AltcoinValidator extends MonetaryValidator {
     @Override
     protected double getMinValue() {
-        return 0.0001;
+        return 0.00000001;
     }
 
     @Override
     protected double getMaxValue() {
-        // Hard to say what the max value should be (zimbabwe dollar....)?
-        // Lets set it to Double.MAX_VALUE until we find some reasonable number
-        return Double.MAX_VALUE;
+        // hard to say what the max value should be with altcoins
+        return 100_000_000;
     }
 
     @Inject
-    public FiatPriceValidator() {
+    public AltcoinValidator() {
     }
 }
