+++ core/src/main/java/bisq/core/util/validation/FiatPriceValidator.java
@@ -15,23 +15,24 @@
  * along with Haveno. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package bisq.desktop.util.validation;
+package bisq.core.util.validation;
 
 import javax.inject.Inject;
 
-public class AltcoinValidator extends MonetaryValidator {
+public class FiatPriceValidator extends MonetaryValidator {
     @Override
     protected double getMinValue() {
-        return 0.00000001;
+        return 0.0001;
     }
 
     @Override
     protected double getMaxValue() {
-        // hard to say what the max value should be with altcoins
-        return 100_000_000;
+        // Hard to say what the max value should be (zimbabwe dollar....)?
+        // Lets set it to Double.MAX_VALUE until we find some reasonable number
+        return Double.MAX_VALUE;
     }
 
     @Inject
-    public AltcoinValidator() {
+    public FiatPriceValidator() {
     }
 }
