+++ desktop/src/main/java/bisq/desktop/main/offer/createoffer/CreateOfferViewModel.java
@@ -20,9 +20,7 @@ package bisq.desktop.main.offer.createoffer;
 import bisq.desktop.Navigation;
 import bisq.desktop.common.model.ViewModel;
 import bisq.desktop.main.offer.MutableOfferViewModel;
-import bisq.desktop.util.validation.AltcoinValidator;
 import bisq.desktop.util.validation.BtcValidator;
-import bisq.desktop.util.validation.FiatPriceValidator;
 import bisq.desktop.util.validation.FiatVolumeValidator;
 import bisq.desktop.util.validation.SecurityDepositValidator;
 
@@ -32,7 +30,8 @@ import bisq.core.provider.price.PriceFeedService;
 import bisq.core.user.Preferences;
 import bisq.core.util.FormattingUtils;
 import bisq.core.util.coin.CoinFormatter;
-
+import bisq.core.util.validation.AltcoinValidator;
+import bisq.core.util.validation.FiatPriceValidator;
 import com.google.inject.Inject;
 
 import javax.inject.Named;
