+++ desktop/src/main/java/bisq/desktop/util/validation/BtcValidator.java
@@ -21,7 +21,7 @@ import bisq.core.btc.wallet.Restrictions;
 import bisq.core.locale.Res;
 import bisq.core.util.FormattingUtils;
 import bisq.core.util.coin.CoinFormatter;
-
+import bisq.core.util.validation.NumberValidator;
 import org.bitcoinj.core.Coin;
 
 import javax.inject.Inject;
