+++ desktop/src/main/java/bisq/desktop/util/validation/FiatVolumeValidator.java
@@ -17,6 +17,7 @@
 
 package bisq.desktop.util.validation;
 
+import bisq.core.util.validation.MonetaryValidator;
 import javax.inject.Inject;
 
 public class FiatVolumeValidator extends MonetaryValidator {
