+++ desktop/src/main/java/bisq/desktop/util/validation/PercentageNumberValidator.java
@@ -18,7 +18,7 @@
 package bisq.desktop.util.validation;
 
 import bisq.core.locale.Res;
-
+import bisq.core.util.validation.NumberValidator;
 import lombok.Setter;
 
 import javax.annotation.Nullable;
