+++ desktop/src/main/java/bisq/desktop/util/validation/SecurityDepositValidator.java
@@ -22,7 +22,7 @@ import bisq.core.locale.Res;
 import bisq.core.payment.PaymentAccount;
 import bisq.core.util.FormattingUtils;
 import bisq.core.util.ParsingUtils;
-
+import bisq.core.util.validation.NumberValidator;
 import javax.inject.Inject;
 
 public class SecurityDepositValidator extends NumberValidator {
